requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
redis>=5.0.0
celery>=5.3.0
paddleocr>=2.7.0
pytesseract>=0.3.10
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.1.0
openai>=1.0.0
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
prometheus-client>=0.19.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
httpx>=0.25.0
aioredis>=2.0.0
pyahocorasick>=2.0.0
llama-index>=0.9.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-embeddings-ollama>=0.1.0
chromadb>=0.4.0
